+++ frontend/riskofbias/robScoreCleanup/components/CheckBoxScoreDisplay.js
@@ -2,10 +2,9 @@ import React, {Component} from "react";
 import PropTypes from "prop-types";
 
 import Study from "study/Study";
-import ScoreBar from "../../../robTable/components/ScoreBar";
-import "./CheckboxScoreDisplay.css";
+import ScoreBar from "../../robTable/components/ScoreBar";
 
-import {SCORE_SHADES, SCORE_TEXT, SCORE_TEXT_DESCRIPTION} from "../../../constants";
+import {SCORE_SHADES, SCORE_TEXT, SCORE_TEXT_DESCRIPTION} from "../../constants";
 
 class CheckboxScoreDisplay extends Component {
     render() {
