--- frontend/riskofbias/robScoreCleanup/components/CheckboxScoreDisplay/CheckboxScoreDisplay.css
@@ -1,11 +1,0 @@
-.score-display-checkbox {
-    margin-top: 15px !important;
-}
-
-.score-display-checkbox p {
-    cursor: pointer;
-}
-
-.score-display-checkbox input {
-    margin: 0 5px 0 0;
-}
