--- frontend/riskofbias/robScoreCleanup/containers/Root/Root.css
@@ -1,13 +1,0 @@
-.cleanStudyMetricForm {
-    border: 2px solid grey;
-    border-radius: 5px;
-    padding: 10px;
-    background-color: #efefef;
-}
-
-.robScoreList {
-    border: 2px solid grey;
-    border-radius: 5px;
-    padding: 10px;
-    background-color: #efefef;
-}
