+++ frontend/riskofbias/robScoreCleanup/containers/ScoreList.js
@@ -4,7 +4,7 @@ import React, {Component} from "react";
 import {inject, observer} from "mobx-react";
 
 import Loading from "shared/components/Loading";
-import CheckboxScoreDisplay from "../../components/CheckboxScoreDisplay";
+import CheckboxScoreDisplay from "../components/CheckboxScoreDisplay";
 
 @inject("store")
 @observer
