+++ frontend/riskofbias/robScoreCleanup/index.js
@@ -3,12 +3,12 @@ import ReactDOM from "react-dom";
 import {Provider} from "mobx-react";
 
 import store from "./store";
-import Root from "./containers/Root";
+import Main from "./containers/Main";
 
 export default function(el) {
     ReactDOM.render(
         <Provider store={store}>
-            <Root />
+            <Main />
         </Provider>,
         el
     );
